Rather than 0 -> A, 1 -> B, try the opposite map
This mapping represents how IBD haplotype codes, which are the integers
zero and one, are translated into their corredponding local ancestry
codes, namely A and B. The plot sept16-10:50.png shows that the local
ancestry does not correlate with IBD, which contradicts the intuition
that IBD and local ancestry should be very correlated. Therefore, it is
reasonable to assume that the haplotype mappings are incorrect. It may
become necessary to perform a much fancier mapping based on the relative
amount of ancestry derived from each group, and picking haplotypes such
that they are the most related. This, however, seems unsound to me,
since we would thus be assuming that what we want to show is already
true!