rewrite data loading to filter entries to HRS AfrAm individuals
Since we're loading Soheil's MergedData dataset, which contains IBD
entries from many other datasets, introducing filtering becomes
necessary to avoid trying to look up the local ancestries of individuals
from datasets other than the AfrAm.

This involves restructuring the way the data is loaded, and since the
MergedData is bigger, it becomes more worthwhile to lazily load it with
generators rather than load it up front, and process it once all the
data is loaded.