exp: final region scope over whole shared region
Right now we're determining whether the shared ancestry is worthwhile
or not by checking what proportion of the chromosome is shared between
the two individuals, and only plotting those who are above a certain
threshold, but this is circular reasoning. With this commit, we pick the
full region, to see whether using the new dataset makes a significant
difference.