add stronger consistency check to plotter
It seems that the bed parser works just fine, so what we'll do is once
we have found the 50 or so longest matches, we'll reload their local
ancestry files and perform a comparison. `to_debugstr` from Individual
will generate a string which is compatible with a bed file, in the
correct order, too, so we'll just compare the outputs for mismatches.