add way to serialize a match object to a string
Loading a match object from a string is more complicated, because we'll only
have partial information concerning the individual's ancestry, so we won't have
the usual list of 22 chromosome objects. This would effectively break any code
that relies on iterating over the chromosomes. Thankfully, there isn't much
practical code that does this. Most times, when this is happening, it's to
perform a test, so the behaviour can easily be changed.